Cleanup main activity
